// src/api/__generated__/story/model/actorCollectionMintDigestResponse.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Mini Drama Service API
 * Mini Drama Platform
 * OpenAPI spec version: v1.0
 */

export interface ActorCollectionMintDigestResponse {
  actorCollectionId?: number;
  assetId?: string;
  creatorAddress?: string;
  walletAddress?: string;
  collectionMintAddress?: string;
  name?: string;
  nftChain?: string;
  nftTokenStandard?: string;
  baseUrl?: string;
  totalSupply?: number;
  initialPriceUsdc?: number;
  trust?: number;
  initialPriceAmount?: string;
  curveMultiplierAmount?: string;
  payToken?: string;
  feeAmount?: string;
  issuedAt?: number;
  expiresAt?: number;
  canonicalPayload?: string;
  sig?: string;
}